+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add support for `:cuprite` and `:playwright` system test drivers
 - Drop support for [End of Life Ruby versions 3.0 and 3.1](https://www.ruby-lang.org/en/downloads/branches/)
 - Drop support for [End of Life Rails versions 6.1, 7.0, and 7.1](https://rubyonrails.org/maintenance)